feat: Setting up a single function for initing simulated backend with all required contracts
